refactor: Step 클래스를 given/when/then 폴더로 분리
- 도메인 기준 → Given/When/Then 기준으로 변경
- Step 검색 시 예측 가능하도록 구조 개선
