+++ src/test/java/com/camping/admin/steps/when/CampsiteWhenSteps.java
@@ -1,36 +1,20 @@
-package com.camping.admin.steps;
+package com.camping.admin.steps.when;
 
-import com.camping.admin.fixture.TestConfig;
 import com.camping.admin.support.AuthHelper;
 import com.camping.admin.support.RestAssuredConfig;
 import com.camping.admin.support.SharedState;
-import io.cucumber.java.en.Given;
 import io.cucumber.java.en.When;
 import io.restassured.RestAssured;
 
 import java.util.Map;
 
-public class CampsiteSteps {
+public class CampsiteWhenSteps {
 
     private final SharedState state;
 
-    public CampsiteSteps(SharedState state) {
+    public CampsiteWhenSteps(SharedState state) {
         this.state = state;
     }
-
-    // === Given ===
-
-    @Given("수정할 캠프사이트가 존재한다")
-    public void 수정할_캠프사이트가_존재한다() {
-        state.setCampsiteId(TestConfig.CampsiteIds.EXISTING);
-    }
-
-    @Given("존재하지 않는 캠프사이트 ID를 사용한다")
-    public void 존재하지_않는_캠프사이트_ID를_사용한다() {
-        state.setCampsiteId(TestConfig.CampsiteIds.NOT_EXIST);
-    }
-
-    // === When ===
 
     @When("캠프사이트 목록을 조회하면")
     public void 캠프사이트_목록을_조회하면() {
