+++ src/test/java/com/camping/admin/steps/when/ProductWhenSteps.java
@@ -1,31 +1,20 @@
-package com.camping.admin.steps;
+package com.camping.admin.steps.when;
 
-import com.camping.admin.fixture.TestConfig;
 import com.camping.admin.support.AuthHelper;
 import com.camping.admin.support.RestAssuredConfig;
 import com.camping.admin.support.SharedState;
-import io.cucumber.java.en.Given;
 import io.cucumber.java.en.When;
 import io.restassured.RestAssured;
 
 import java.util.Map;
 
-public class ProductSteps {
+public class ProductWhenSteps {
 
     private final SharedState state;
 
-    public ProductSteps(SharedState state) {
+    public ProductWhenSteps(SharedState state) {
         this.state = state;
     }
-
-    // === Given ===
-
-    @Given("수정할 상품이 존재한다")
-    public void 수정할_상품이_존재한다() {
-        state.setProductId(TestConfig.ProductIds.LANTERN);
-    }
-
-    // === When ===
 
     @When("상품을 생성하면")
     public void 상품을_생성하면() {
